feat: del global decorators & add button component decorators
